+++ AppSystemAgenceBack/src/main/java/com/intiformation/appAgenceImmo/modele/AgentImmobilier.java
@@ -26,13 +26,13 @@ public class AgentImmobilier extends Personne {
 	}
 	
 	// ctor ALL PROPS 
-	public AgentImmobilier(int id_personne, String nom, String prenom, String email, Adresse adresse, String motDePasse) {
+	public AgentImmobilier(Long id_personne, String nom, String prenom, String email, Adresse adresse, String motDePasse) {
 		super(id_personne, nom, prenom, email, adresse);
 		this.motDePasse = motDePasse;
 	}
 
 	// ctor SANS Adresse 
-	public AgentImmobilier(int id_personne, String nom, String prenom, String email, String motDePasse) {
+	public AgentImmobilier(Long id_personne, String nom, String prenom, String email, String motDePasse) {
 		super(id_personne, nom, prenom, email);
 		this.motDePasse = motDePasse;
 	}
